% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_markers.R
\name{add_markers}
\alias{add_markers}
\title{Returns markers between selected celltypes}
\usage{
add_markers(sce, celltypes = NULL, Top = 1)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} object representing scRNA-seq counts matrix containing celltype field.}

\item{celltypes}{Celltypes for which we want to find markers. Default=NULL and in this case all celltypes will be included.}

\item{Top}{Positive integer representing how many Top markers should be returned.}
}
\value{
A \code{data.frame} object with selected markers.
}
\description{
Returns markers between selected celltypes
}
\examples{
require(SingleCellExperiment)
n_row = 30000
n_col = 100
sce = SingleCellExperiment(assays = list(logcounts = matrix(rnorm(n_row*n_col), ncol=n_col)))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
sce$celltype = as.factor(sample(1:10, n_col, replace=TRUE))
out = add_markers(sce)
}
