% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_rare_celltypes.R
\name{detect_rare_celltypes}
\alias{detect_rare_celltypes}
\title{Detecting and filtering rare celltypes from count matrix. Additionally, assigning markers for rare celltypes from the original count matrix.}
\usage{
detect_rare_celltypes(
  sce,
  n_cells.thresh = 10,
  Top = 5,
  filter_rare_celltypes = TRUE,
  get_markers_rare_celltypes = TRUE
)
}
\arguments{
\item{sce}{SingleCellExperiment object representing scRNA-seq counts matrix containing celltype field.}

\item{n_cells.thresh}{Threshold of how many cells should be used to detect rare celltypes. Default=10.}

\item{Top}{Positive integer representing how many Top markers should be returned.}

\item{filter_rare_celltypes}{Boolean representing whether filtered counts matrix should be returned.}

\item{get_markers_rare_celltypes}{Boolean representing whether markers for rare celltypes should be returned.}
}
\value{
Filtered counts matrix and/or markers for rare celltypes.
}
\description{
Detecting and filtering rare celltypes from count matrix. Additionally, assigning markers for rare celltypes from the original count matrix.
}
