% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ct_hierarchy.R
\name{.get_listed_ct_hierarchy}
\alias{.get_listed_ct_hierarchy}
\title{Transforms celltype hierarchy representation to the nested list format.}
\usage{
.get_listed_ct_hierarchy(ct_hierarchy)
}
\arguments{
\item{ct_hierarchy}{hclust object representing celltype hierarchy.}
}
\value{
Celltype hierarchy in the nested lists format.
}
\description{
Transforms celltype hierarchy representation to the nested list format.
}
