% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{.get_mapping}
\alias{.get_mapping}
\title{.get_mapping}
\usage{
.get_mapping(
  sce,
  genes = rownames(sce),
  batch = NULL,
  n.neigh = 5,
  nPC = 50,
  get.dist = F,
  cosine = F
)
}
\arguments{
\item{sce}{SingleCellExperiment object containing gene counts matrix (stored in 'logcounts' assay).}

\item{genes}{Character vector containing genes that are used to construct kNN-graph.}

\item{batch}{Name of the field in colData(sce) to specify batch. Default batch=NULL if no batch is applied.}

\item{n.neigh}{Scalar specifying number of neighbors to use for kNN-graph. Default n.neigh=5.}

\item{nPC}{Scalar (or NULL, if no PCA applied) specifying number of PCs to use for kNN-graph. Default nPC=50.}

\item{get.dist}{Boolean specifying if distances for kNN-graph should be returned. Default get.dist=FALSE.}

\item{cosine}{Boolean specifying if cosine normalization should be applied prior to constructing kNN-graph. Default cosine=FALSE.}
}
\value{
List containing entry named 'cells_mapped': data.frame, each row corresponds to cellID (in rownames) and columns contain information about first K neighbors;
(if get.dist=T) entry named 'distances': corresponding to 'cells_mapped' data.frame, columns contain distances to the first K neighbors.
}
\description{
For each cell, function assigns neighbor cells (in kNN-graph) and corresponding distances.
}
