% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{.get_MNN_corrected_mapping}
\alias{.get_MNN_corrected_mapping}
\title{.get_MNN_corrected_mapping}
\usage{
.get_MNN_corrected_mapping(
  sce,
  genes = rownames(sce),
  batch = NULL,
  n.neigh = 5,
  nPC = 50,
  cosine = F
)
}
\arguments{
\item{sce}{SingleCellExperiment object containing gene counts matrix (stored in 'logcounts' assay).}

\item{genes}{Character vector containing genes that are used to construct kNN-graph.}

\item{batch}{Name of the field in colData(sce) to specify batch. Default batch=NULL if no batch is applied.}

\item{n.neigh}{Scalar specifying number of neighbors to use for kNN-graph. Default n.neigh=5.}

\item{nPC}{Scalar specifying number of PCs to use for kNN-graph. Default nPC=50.}

\item{cosine}{Boolean specifying if cosine normalization should be applied prior to constructing kNN-graph. Default cosine=FALSE.}
}
\description{
For each cell, function assigns neighbor cells (in MNN-correced kNN-graph) and corresponding distances.
}
