% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_highly_expressed_genes.R
\name{filter_highly_expressed_genes}
\alias{filter_highly_expressed_genes}
\title{Filtering for highly expressed genes.}
\usage{
filter_highly_expressed_genes(sce, method = "absolute", logcounts.thresh = 20)
}
\arguments{
\item{sce}{SingleCellExperiment object representing scRNA-seq counts matrix containing celltype field.}

\item{method}{String specifying the approach to take to detect highly expressed genes. If set to "absolute", all genes that have q75 higher than a threshold specified in logcounts.thresh (in at least one celltype) will be discarded. If set to "ribosome_based", the logcounts.thresh is calculated as minimum of upper bounds of q75 across all stably expressed ribosomal genes.}

\item{logcounts.thresh}{Numeric positive scalar specifying the threshold to detect highly expressed genes.}
}
\value{
Filtered counts matrix (SingleCellExperiment object).
}
\description{
Filtering for highly expressed genes.
}
\examples{
require(SingleCellExperiment)
n_row = 30000
n_col = 100
sce = SingleCellExperiment(assays = list(logcounts = matrix(rnorm(n_row*n_col), ncol=n_col)))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
sce$celltype = as.factor(sample(1:5, n_col, replace=TRUE))
sce = filter_highly_expressed_genes(sce)
sce = filter_highly_expressed_genes(sce, "ribosome_based")

}
