% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_celltype_mapping.R
\name{get_celltype_mapping}
\alias{get_celltype_mapping}
\title{get_celltype_mapping}
\usage{
get_celltype_mapping(
  sce,
  genes.selection,
  batch = NULL,
  n.neigh = 5,
  nPC.selection = NULL,
  cosine = F,
  return.stat = T,
  which_genes_to_use = "all",
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment object containing gene counts matrix (stored in 'logcounts' assay).}

\item{genes.selection}{Character vector specifying genes to be used for the construction of Selection kNN-graph.}

\item{batch}{Name of the field in colData(sce) to specify batch. Default batch=NULL if no batch is applied.}

\item{n.neigh}{Positive integer > 1, specifying number of neighbors to use for kNN-graph. Default n.neigh=5.}

\item{nPC.selection}{Scalar specifying number of PCs to use for construction of True kNN-graph. Default nPC=NULL.}

\item{cosine}{Boolean specifying if cosine normalization should be applied prior to constructing kNN-graph. Default cosine=FALSE.}

\item{return.stat}{Boolean specifying if stat on the mapping should be returned alongside the mapping itself.}

\item{which_genes_to_use}{String specifying whether celltype mapping should be performed only on DE (between celltypes) genes (= 'DE') or all inserted genes (= 'all'). Default option="all".}

\item{...}{Additional arguments (e.g. the ones you can pass to get_DE_genes).}
}
\value{

}
\description{
For each cell, returns actual and mapped (most common among neighbors in the Selection graph) celltype labels.
}
\examples{
require(SingleCellExperiment)
n_row = 1000
n_col = 100
sce = SingleCellExperiment(assays = list(logcounts = matrix(rnorm(n_row*n_col), ncol=n_col)))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
sce$celltype = as.character(sample.int(5, n_col, replace = TRUE))
genes.selection = sample( rownames(sce) , 20)
out = get_celltype_mapping(sce, genes.selection = genes.selection)

}
