% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ct_hierarchy.R
\name{get_ct_hierarchy}
\alias{get_ct_hierarchy}
\title{Produces binary tree representing celltype relationships (based on input genes)}
\usage{
get_ct_hierarchy(
  sce,
  genes = NULL,
  batch = NULL,
  cosineNorm = TRUE,
  nPC = NULL,
  p.thresh = 0.05
)
}
\arguments{
\item{sce}{SingleCellExperiment object representing scRNA-seq counts matrix containing celltype field.}

\item{genes}{Genes that are used to assess celltypes similarity. Default=NULL, and in this case all rownames of sce would be assigned as genes.}

\item{batch}{A field from colData of sce object that indicates which field is used as batch. If specified, PCA would be calculated using multiBatchPCA and pcs would be batch-corrected. Default=NULL that assumes no batch.}

\item{cosineNorm}{A boolean specifying whether cosine normalization should be applied to logcounts. Default=TRUE.}

\item{nPC}{Number of PCs to be used. Default=NULL, and in this case all significant PCs (reflecting dependence between celltype and PC from ANOVA) will be included.}

\item{p.thresh}{A P-value threshold to determine significance of each PC. Default=0.05.}
}
\value{
hclust object representing relationship of celltypes.
}
\description{
Produces binary tree representing celltype relationships (based on input genes)
}
\examples{
require(SingleCellExperiment)
n_row = 30000
n_col = 100
sce = SingleCellExperiment(assays = list(logcounts = matrix(rnorm(n_row*n_col), ncol=n_col)))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
sce$celltype = as.factor(sample(1:5, n_col, replace=TRUE))
ct_hierarchy = get_ct_hierarchy(sce , genes = c(1:20), nPC = 10)

}
