% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DE_genes.R
\name{get_DE_genes}
\alias{get_DE_genes}
\title{Returns all genes that are deferentially expressed in at least one celltype}
\usage{
get_DE_genes(
  sce,
  test = "binom",
  FDR.thresh = 0.01,
  filter_by_ct_specificity = TRUE
)
}
\arguments{
\item{sce}{SingleCellExperiment object representing scRNA-seq counts matrix containing celltype field}

\item{test}{A parameter of findMarkers function determining which statistical test should be used for selection of DE genes. Default=binom.}

\item{FDR.thresh}{A False Discovery Rate threshold determining whether given gene would be considered as DE. Default=0.01}

\item{filter_by_ct_specificity}{A boolean parameter to whether we should perform initial filtering by gene expression being CT specific (q75 == 0 in more than half of celltypes)}
}
\value{
data.frame object where each row represents gene/celltype if the gene is assigned as DE in a given celltype
}
\description{
Returns all genes that are deferentially expressed in at least one celltype
}
\examples{
require(SingleCellExperiment)
n_row = 30000
n_col = 100
sce = SingleCellExperiment(assays = list(logcounts = matrix(rnorm(n_row*n_col), ncol=n_col)))
rownames(sce) = as.factor(1:n_row)
sce$celltype = as.factor(sample(1:5, n_col, replace=TRUE))

markers.binom = get_DE_genes(sce)
markers.t = get_DE_genes(sce , test = "t")

}
