% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_prediction_scores.R
\name{get_gene_correlation_scores}
\alias{get_gene_correlation_scores}
\title{get_gene_correlation_scores}
\usage{
get_gene_correlation_scores(
  sce,
  genes,
  batch = NULL,
  n.neigh = 5,
  nPC = NULL,
  genes.predict = rownames(sce),
  method = "spearman",
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment object containing gene counts matrix (stored in 'logcounts' assay).}

\item{genes}{Character vector containing names of selected genes.}

\item{batch}{Name of the field in colData(sce) to specify batch. Default batch=NULL if no batch is applied.}

\item{n.neigh}{Positive integer > 1, specifying number of neighbors to use for kNN-graph. Default n.neigh=5.}

\item{nPC}{Scalar (or NULL) specifying number of PCs to use for construction of kNN-graph. Default nPC=NULL.}

\item{genes.predict}{Character vector containing names of genes for which we want to calculate gene prediction score. Default = rownames(sce).}

\item{method}{Character specifying method for correlation. Availbale options are c("spearman", "pearson", "kendall"). Default method="spearman".}

\item{...}{Additional arguments.}
}
\value{

}
\description{
Calculates per gene correlation between measured expression levels and estimated expression levels (average across neighbors).
It is an intermediate step for 'get_gene_prediction_scores' function. It can be handy to calculate this part separately to be recycled multiple times
to compare different seelctions.
}
\examples{
require(SingleCellExperiment)
n_row = 1000
n_col = 100
sce = SingleCellExperiment(assays = list(logcounts = matrix(rnorm(n_row*n_col), ncol=n_col)))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
genes = rownames(sce)
out = get_gene_correlation_scores(sce, genes)

}
