% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_ranking.R
\name{get_gene_ranking}
\alias{get_gene_ranking}
\title{Returns ordered gene selection (by their relevance to PC-space)}
\usage{
get_gene_ranking(sce, genes, type = "none")
}
\arguments{
\item{sce}{SingleCellExperiment object representing scRNA-seq counts matrix}

\item{genes}{Initially pre-selected DE (per celltype) genes that we assign as relevant for generation of PC-space}

\item{type}{Setting to assign whether covariance matrix in PCA should be reweighted so each celltype has the same weight (celltype). Otherwise setting should be set to none and each cell will be accounted once.}
}
\value{
list containing stat per PC and order gene vector
}
\description{
Returns ordered gene selection (by their relevance to PC-space)
}
\examples{
require(SingleCellExperiment)
n_row = 30000
n_col = 100
sce = SingleCellExperiment(assays = list(logcounts = matrix(rnorm(n_row*n_col), ncol=n_col)))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
sce$celltype = as.factor(sample(1:5, n_col, replace=TRUE))
genes_ranked = get_gene_ranking(sce , genes = c(1:20))
}
