% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neighborhood_preservation_scores.R
\name{get_neighborhood_preservation_scores}
\alias{get_neighborhood_preservation_scores}
\title{get_neighborhood_preservation_scores}
\usage{
get_neighborhood_preservation_scores(
  sce,
  neighs.all = NULL,
  genes.all = rownames(sce),
  genes.selection,
  batch = NULL,
  n.neigh = 5,
  nPC.all = 50,
  nPC.selection = NULL,
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment object containing gene counts matrix (stored in 'logcounts' assay).}

\item{neighs.all}{If not NULL, contains information about True kNN-graph (for each cell - ordered neighbors and distances). Useful to have a priori if cell score will be recalculated multiple times.}

\item{genes.all}{String specifying genes to be used for construction of True kNN-graph.}

\item{genes.selection}{String specifying genes to be used for construction of Selection kNN-graph.}

\item{batch}{Name of the field in colData(sce) to specify batch. Default batch=NULL if no batch is applied.}

\item{n.neigh}{Scalar specifying number of neighbors to use for kNN-graph. Default n.neigh=5.}

\item{nPC.all}{Scalar specifying number of PCs to use for construction of True kNN-graph. Default nPC.all=50.}

\item{nPC.selection}{Scalar specifying number of PCs to use for construction of True kNN-graph. Default nPC.selection=NULL (no PCA).}

\item{...}{Additional arguments}
}
\value{
data.frame, each row corresponds to cell from counts matrix, contains field cell_score = cell neighborhood preservation score
}
\description{
Calculates cell neighborhood preservation scores by comparing neighbors from True and Selection k-NN graphs.
}
\examples{
require(SingleCellExperiment)
n_row = 1000
n_col = 100
sce = SingleCellExperiment(assays = list(logcounts = matrix(rnorm(n_row*n_col), ncol=n_col)))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
genes.selection = sample(rownames(sce) , 20)
out = get_neighborhood_preservation_scores(sce, genes.selection = genes.selection)

}
