% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neighborhood_preservation_scores.R,
%   R/get_neighborhood_preservation_scores_alternative.R
\name{get_neighs_all_stat}
\alias{get_neighs_all_stat}
\title{get_neighs.all_stat}
\usage{
get_neighs_all_stat(
  sce,
  genes.all = rownames(sce),
  batch = NULL,
  n.neigh = 5,
  nPC.all = 50
)

get_neighs_all_stat(
  sce,
  genes.all = rownames(sce),
  batch = NULL,
  n.neigh = 5,
  nPC.all = 50
)
}
\arguments{
\item{sce}{SingleCellExperiment object containing gene counts matrix.}

\item{genes.all}{String specifying genes to be used for construction of True kNN-graph.}

\item{batch}{Name of the field in colData(sce) to specify batch. Default batch=NULL if no batch is applied.}

\item{n.neigh}{Positive integer > 1, specifying number of neighbors to use for kNN-graph. Default n.neigh=5.}

\item{nPC.all}{Scalar specifying number of PCs to use for construction of True kNN-graph. Default nPC.all=50.}

\item{...}{Additional arguments.}
}
\value{
kNN-graph with assigned neighbors and corresponding z-scored distances.

kNN-graph with assigned neighbors and corresponding z-scored distances.
}
\description{
get_neighs.all_stat

get_neighs.all_stat
}
