% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_redundancy_stat.R
\name{get_redundancy_stat}
\alias{get_redundancy_stat}
\title{Functions calculates relevance of each gene (within current selection) to celltype mapping.}
\usage{
get_redundancy_stat(sce, genes, genes_to_assess = genes, batch = NULL, ...)
}
\arguments{
\item{sce}{SingleCellExperiment object containing gene counts matrix (stored in 'logcounts' assay).}

\item{genes}{Character vector specifying selected library.}

\item{genes_to_assess}{Character vector specifying gene names for which we want to assess redundancy. Should be a subset of genes.}

\item{batch}{Name of the field in colData(sce) to specify batch. Default batch=NULL if no batch is applied.}

\item{...}{Additional arguments you can pass to get_celltype_mapping.}
}
\value{
data.frame, each row corresponds to celltype/gene; frac_correctly_mapped_ratio corresponds to ratio between accuracy of the mapping for a given celltype while using genes excluding and including the gene in question.
}
\description{
Functions calculates relevance of each gene (within current selection) to celltype mapping.
}
\examples{
require(SingleCellExperiment)
n_row = 1000
n_col = 100
sce = SingleCellExperiment(assays = list(logcounts = matrix(rnorm(n_row*n_col), ncol=n_col)))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
sce$celltype = as.character(sample.int(5, n_col, replace = TRUE))
genes = rownames(sce)
genes_to_assess = sample(rownames(sce),5)
out = get_redundancy_stat(sce, genes, genes_to_assess = genes_to_assess)

}
