% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rid_of_rare_celltypes.R
\name{get_rid_of_rare_celltypes}
\alias{get_rid_of_rare_celltypes}
\title{Filters rare celltypes}
\usage{
get_rid_of_rare_celltypes(sce, n_cells.thresh = 10)
}
\arguments{
\item{sce}{\itemize{
\item a SingleCellExperiment object that contains celltype field
}}

\item{n_cells.thresh}{\itemize{
\item a threshold of how many cells should be used to detect rare celltypes
}}
}
\value{
filtered sce
}
\description{
Filters rare celltypes
}
\examples{

}
