% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_mapping.R
\name{get_sensitivity_mapping}
\alias{get_sensitivity_mapping}
\title{Returns celltype corresponding vector representing percentage of cells from the celltype that is mapped correctly}
\usage{
get_sensitivity_mapping(mapping)
}
\arguments{
\item{mapping}{A \code{data.frame} object representing mapping.}
}
\value{
A \code{data.frame} object with fields 'celltype' and 'frac_mapped_correctly'.
}
\description{
Returns celltype corresponding vector representing percentage of cells from the celltype that is mapped correctly
}
\examples{
require(SingleCellExperiment)
n_row = 30000
n_col = 100
cells_reference = c(1:80)
sce = SingleCellExperiment(assays = list(logcounts = matrix(rnorm(n_row*n_col), ncol=n_col)))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
sce$celltype = as.factor(sample(1:5, n_col, replace=TRUE))
genes = c(1:20)
sce_reference = sce[ , colnames(sce) \%in\% cells_reference]
sce_query = sce[ , !colnames(sce) \%in\% cells_reference]
out = hierarchical_mapping(sce_reference , sce_query , genes, nPC = 5, nPC.ct_hierarchy=5)
mapping = out$mapping
sensitivity_stat = get_sensitivity_mapping(mapping)
}
