% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_umap_coordinates.R
\name{get_umap_coordinates}
\alias{get_umap_coordinates}
\title{get_umap_coordinates}
\usage{
get_umap_coordinates(sce, genes, batch = NULL, nPC = length(genes) - 1)
}
\arguments{
\item{sce}{SingleCellExperiment object containing gene counts matrix (stored in 'logcounts' assay).}

\item{genes}{Character vector specifying genes to use for calculating umap-coordinates.}

\item{batch}{Name of the field in colData(sce) to specify batch. Default batch=NULL if no batch is applied.}

\item{nPC}{Scalar specifying number of PCs to use. Default = length(genes)-1}
}
\value{
Return data.frame, where UMAP-coordinates are assigned as x and y.
}
\description{
Function returns UMAP-coordinates if onle selected genes are used.
}
\examples{
require(SingleCellExperiment)
n_row = 1000
n_col = 100
sce = SingleCellExperiment(assays = list(logcounts = matrix(rnorm(n_row*n_col), ncol=n_col)))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
out = get_umap_coordinates(sce, genes = rownames(sce)[1:10])

}
