% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping}
\alias{mapping}
\title{MNN-based mapping of celltypes}
\usage{
mapping(
  sce_reference,
  sce_query,
  genes,
  batch = NULL,
  cosineNorm = TRUE,
  n.neigh = 10,
  nPC = NULL,
  p.thresh = 0.05
)
}
\arguments{
\item{sce_reference}{SingleCellExperiment object representing reference scRNA-seq counts matrix containing celltype field.}

\item{sce_query}{SingleCellExperiment object representing query scRNA-seq counts matrix containing celltype field. Cells and accordingly celltype labels from this ds would be assigned neighbors from sce_reference.}

\item{genes}{Genes that are used for integration between sce_reference and sce_query}

\item{batch}{A field from colData of sce object that indicates which field is used as batch. If specified, PCA would be calculated using multiBatchPCA and pcs would be batch-corrected. Default=NULL that assumes no batch.}

\item{cosineNorm}{A boolean specifying whether cosine normalization should be applied to logcounts. Default=TRUE.}

\item{n.neigh}{Number of nearets neighbrs used for the integration. Default=10.}

\item{nPC}{Number of PCs to be used. Default=NULL, and in this case all significant PCs (reflecting dependence between celltype and PC from ANOVA) will be included.}

\item{p.thresh}{A P-value threshold to determine significance of each PC. Default=0.05.}
}
\value{
data.frame object, where every row corresponds to cell from sce_query. celltype_mapped field contains assigned from sce_reference celltype.
}
\description{
MNN-based mapping of celltypes
}
\examples{
require(SingleCellExperiment)
n_row = 30000
n_col = 100
cells_reference = c(1:80)
sce = SingleCellExperiment(assays = list(logcounts = matrix(rnorm(n_row*n_col), ncol=n_col)))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
sce$celltype = as.factor(sample(1:5, n_col, replace=TRUE))
sce_reference = sce[ , colnames(sce) \%in\% cells_reference]
sce_query = sce[ , !colnames(sce) \%in\% cells_reference]
genes = c(1:20)
out = mapping(sce_reference , sce_query , genes, nPC=5)
}
