% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_expr_distribution}
\alias{plot_expr_distribution}
\title{Distribution of gene expression per celltype}
\usage{
plot_expr_distribution(sce, gene, assay = "logcounts", title = gene)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} object representing scRNA-seq counts matrix containing celltype field.}

\item{gene}{Gene which expression we want to plot.}

\item{assay}{A string specifying which assay from sce to use. Default=logcounts.}

\item{title}{A string specifying title. Default=gene.}
}
\value{
A \code{ggplot} object containing boxplots per celltype.
}
\description{
Distribution of gene expression per celltype
}
