% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_expression_heatmap}
\alias{plot_expression_heatmap}
\title{plot_expression_heatmap}
\usage{
plot_expression_heatmap(sce, genes, value.type)
}
\arguments{
\item{sce}{SingleCellExperiment object containing gene counts matrix (stored in 'logcounts' assay).}

\item{genes}{Character vector containing gene names.}

\item{value.type}{String specifying whether to plot average expression (= "mean") or fraction of cells with non-zero counts(= "frac").}
}
\value{
Heatmap with average across (per gene/per celltype)
}
\description{
plot_expression_heatmap
}
