% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_mapping_heatmap}
\alias{plot_mapping_heatmap}
\title{Heatmap representing how well celltypes are mapped}
\usage{
plot_mapping_heatmap(mapping, title = NULL)
}
\arguments{
\item{mapping}{A \code{data.frame} object representing mapping.}

\item{title}{A string specifying title. Default=gene.}
}
\value{
A \code{ggplot} object containing boxplots per celltype. X-axis corresponds to actual celltype, Y-axis corresponds to mapped celltypes, and color gradient corresponds to fraction of cells from actual celltype that are mapped to mapped celltype.
}
\description{
Heatmap representing how well celltypes are mapped
}
