% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_redundancy_stat}
\alias{plot_redundancy_stat}
\title{plot_redundancy_stat}
\usage{
plot_redundancy_stat(
  redundancy_stat,
  celltypes = unique(redundancy_stat$celltype),
  genes = unique(redundancy_stat$gene)
)
}
\arguments{
\item{redundancy_stat}{data.frame - an output of get_redundancy_stat}

\item{celltypes}{Character vector specifying for which celltypes the output should be plotted. Default celltypes = unique(redundancy_stat$celltype)}

\item{genes}{Character vector specifying for which genes the output should be plotted. Default genes = unique(redundancy_stat$gene)}
}
\value{
Heatmap of redundancy in celltype mapping
}
\description{
plot_redundancy_stat
}
