% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_umaps_w_counts}
\alias{plot_umaps_w_counts}
\title{plot_umaps_w_counts}
\usage{
plot_umaps_w_counts(sce, genes, size = 0.25, ncol = NULL)
}
\arguments{
\item{sce}{SingleCellExperiment object containing gene counts matrix (stored in 'logcounts' assay). Should contain field 'UMAP' in reducedDim slot, colnames = x and y.}

\item{genes}{Character vector containing gene names for which to plot UMAPs.}

\item{size}{Size of dots for geom_point}

\item{ncol}{Positive integer specifying number of columns for ggarrange}
}
\value{
For each gene: scater plot with UMAP-coordinates colored by gene expression
}
\description{
plot_umaps_w_counts
}
