% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retain_informative_genes.R
\name{retain_informative_genes}
\alias{retain_informative_genes}
\title{retain_informative_genes}
\usage{
retain_informative_genes(sce, n = NULL, var.thresh = 0, discard.mt = T)
}
\arguments{
\item{sce}{SingleCellExperiment object containing gene counts matrix (stored in 'logcounts' assay).}

\item{n}{Scalar (or NULL, if not applied) specifying number of top HVGs to be return. Default=NULL.}

\item{var.thresh}{Scalar (float) specifying threshold for variation to filter HVGs. Default=0.}

\item{discard.mt}{Boolean specifying if mitochondrial genes should be discarded. Default=TRUE.}
}
\value{
Reduced sce with prefiltered set of genes.
}
\description{
Function selects variable genes and retains them to represent counts matrix for downstream analysis. This is essentially a wrapper for scran::getTopHVGs +
simple check for mt-genes.
}
\examples{
require(SingleCellExperiment)
n_row = 1000
n_col = 100
sce = SingleCellExperiment(assays = list(logcounts = matrix(rnorm(n_row*n_col), ncol=n_col)))
rownames(sce) = as.factor(1:n_row)
colnames(sce) = c(1:n_col)
sce$cell = colnames(sce)
out = retain_informative_genes(sce, n = 100)

}
