% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_mouseEmbryo.R
\docType{data}
\name{sce_mouseEmbryo}
\alias{sce_mouseEmbryo}
\title{MouseEmbryo, E8.5
Counts matrix for mouse embryo, E8.5 (data extracted from Pijuan-Sala et al., 2019).}
\format{
SCE object
}
\source{
MouseGastrulationData package
}
\usage{
data(sce_mouseEmbryo)
}
\description{
SCE object with log-normalised counts matrix (pre-calculated with custom script) strored in assay logcounts.
colData contains information about sample (i.e. batch), celltype and provided color-code for celltypes.
reducedDim contains UMAP coordinates.
Counts matrix is already pre-processed and only non-mitochondrial genes with variance > 0 are retained (top 3000 to reduce memory usage and computational time).
}
