% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_mapping.R
\name{get_sensitivity_mapping}
\alias{get_sensitivity_mapping}
\title{Returns celltype corresponding vector representing percentage of cells from the celltype that is mapped correctly}
\usage{
get_sensitivity_mapping(mapping)
}
\arguments{
\item{mapping}{A \code{data.frame} object representing mapping.}
}
\value{
A \code{data.frame} object with fields 'celltype' and 'frac_mapped_correctly'.
}
\description{
Returns celltype corresponding vector representing percentage of cells from the celltype that is mapped correctly
}
