% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_mapping_heatmap}
\alias{plot_mapping_heatmap}
\title{plot_mapping_heatmap}
\usage{
plot_mapping_heatmap(
  mapping,
  levels = unique(mapping$celltype),
  title = NULL,
  ...
)
}
\arguments{
\item{mapping}{data.frame containing celltype mapping information.}

\item{levels}{Character vector specifying the order for plotting.}

\item{title}{String to be passed as a title. Default title=NULL.}

\item{...}{Additional arguments to pass.}
}
\value{
Heatmap for confusion matrix.
}
\description{
plot_mapping_heatmap
}
