% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_to_sce.R
\name{raw_to_sce}
\alias{raw_to_sce}
\title{raw_to_sce}
\usage{
raw_to_sce(
  counts_dir,
  counts_type = "counts",
  transform_counts_to_logcounts = TRUE,
  header = TRUE,
  sep = "\\t",
  meta_dir = NULL,
  batch = NULL,
  ...
)
}
\arguments{
\item{counts_dir}{String specifying the directory for counts matrix (assuming counts where already calculated)}

\item{counts_type}{String specifying whether raw data is stored as counts or log-counts. For geneBasis we recommend to work with log-counts.}

\item{transform_counts_to_logcounts}{In case, raw data are counts, Boolean specifying whether we should perform log-normalization.}

\item{header}{Boolean specifying if logcounts_dir file has cell IDs stored in colnames.}

\item{sep}{the field separator string. Note it should be the same for logcounts_dir and meta_dir (if exists)}

\item{meta_dir}{If not NULL, a string specifying the directory for meta-data (i.e. celltype, batch, UMAP-coordinates).}

\item{batch}{If not NULL (no batch applied), string specifying a column in meta file that will be used as batchID. Please check that named column exists in meta-file.}

\item{...}{Additional arguments.
Note that UMAP coordinates (if passed in meta-data) should be named as x and y. Also, if meta contains field cell - this field will be used for cell IDs.}
}
\value{
SingleCellExperiment object with gene counts/logcounts and meta-data (if applicable)
}
\description{
Simply a parser raw .txt files --> SingleCellExperiment object + adding logcounts
}
\examples{
require(SingleCellExperiment)
counts_dir = system.file("extdata", "raw_spleen.txt", package = "geneBasisR")
meta_dir = system.file("extdata", "raw_spleen_meta.txt", package = "geneBasisR")
out = raw_to_sce(counts_dir, counts_type = "logcounts", transform_counts_to_logcounts = FALSE, header = TRUE, sep = "\t" , meta_dir = meta_dir, batch = NULL)

}
