% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_to_sce.R
\name{raw_to_sce}
\alias{raw_to_sce}
\title{raw_to_sce}
\usage{
raw_to_sce(
  counts_dir,
  counts_type = "counts",
  transform_counts_to_logcounts = TRUE,
  header = TRUE,
  sep = "\\t",
  meta_dir = NULL,
  batch = NULL,
  ...
)
}
\arguments{
\item{counts_dir}{String specifying the directory for counts matrix (assuming counts where already calculated)}

\item{counts_type}{String specifying whether raw data is stored as counts or log-counts (= 'counts' and 'logcounts' respectively).
For geneBasis we recommend to work with log-counts. If you do not have log-counts precomputed, they can be computed within this function.}

\item{transform_counts_to_logcounts}{In case, raw data are counts (as opposed to log-counts), Boolean specifying whether we should perform log-normalization.}

\item{header}{Boolean specifying if logcounts_dir file has cell IDs stored in colnames.}

\item{sep}{the field separator string. Note it should be the same for logcounts_dir and meta_dir (if latter exists).}

\item{meta_dir}{If not NULL, a string specifying the directory for meta-data (i.e. celltype, batch, UMAP-coordinates).
Store UMAP-coordinates as 'x' and 'y' (relevant for plotting functions).
Also, if meta contains field cell - this field will be used for cell IDs (so ensure the values are unique).}

\item{batch}{If not NULL (i.e. no batch), string specifying a column in meta file that will be used as batchID. Please check that specified batch name exists in meta-file.}

\item{...}{Additional arguments. This includes d and min.mean for scran::quickCluster - used to calculate size factors to compute normalized log-counts.}
}
\value{
SingleCellExperiment object with gene counts/logcounts and meta-data (if supplied) stored in colData.
}
\description{
Essentially a parser for raw .txt counts file into SingleCellExperiment object of the right format. If raw counts are passed, log-normalization is performed
(optional but recommended) and logcounts will be used downstream.
}
\examples{
require(SingleCellExperiment)
counts_dir = system.file("extdata", "raw_spleen.txt", package = "geneBasisR")
meta_dir = system.file("extdata", "raw_spleen_meta.txt", package = "geneBasisR")
out = raw_to_sce(counts_dir, counts_type = "logcounts", transform_counts_to_logcounts = FALSE, header = TRUE, sep = "\t" , meta_dir = meta_dir, batch = NULL)

}
